Add versions of benchmarks without input preprocessing
Summary: Add versions of the ATen, PT2, and CuteDSL benchmarks that don't include input preprocessing. This way, we can choose to time raw kernel performance. In some real world scenarios, we may receive inputs that are already in the correct format (or in the CuteDSL scenario, we may have already cached the compiled kernel).

Differential Revision: D82673867