Fix setting data partition factor in FA
With https://github.com/facebookexperimental/triton/pull/458, the data
partition factor is now controlled by an argument to tl.range rather
than by the config.

Note that this requires the ws-3.5 branch to be able to run the
benchmark now, which does not seem ideal. I experimented with a few ways
to make it work with main but could not find a clean solution.