Add "tagging" benchmark
Summary:
Read the AST of the target op and search through call sites to find whether it is linked to a Triton op or native custom op.

Requires the ability to import the operator classes.

Note this will only work for a few easy operators like embedding and ragged_attention. Need more work to get it done for most of the ops.

Differential Revision: D86154623